+++ Changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Add `MmioSerialPort::new_with_stride` function ([#36](https://github.com/rust-osdev/uart_16550/pull/36))
+
 # 0.3.1 – 2024-07-11
 
 - Add `try_send_raw` and `try_receive` ([#34](https://github.com/rust-osdev/uart_16550/pull/34))